Adds default main and test main methods
In the future I'd like to determine a system by which users can pass in a template to use. This would make the app that much more usable. The idea being that you could pass in some params or config data that would allow users to utilze their own templates for these main methods